# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Post-Meeting Social Media Generator API
  description: |
    Enterprise-grade API for generating social media content from meeting transcripts.
    
    ## Features
    - Google Calendar integration
    - Recall.ai meeting recording
    - AI-powered content generation
    - Social media posting
    - Advanced analytics
    - Enterprise integrations
    
    ## Authentication
    All API endpoints require authentication via NextAuth.js session tokens.
    
    ## Rate Limiting
    API requests are rate limited per user:
    - 1000 requests per hour for authenticated users
    - 100 requests per hour for unauthenticated users
    
    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: API Support
    email: support@postmeeting.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.postmeeting.com/v1
    description: Production server
  - url: https://staging-api.postmeeting.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
paths:
  /meetings:
    get:
      summary: List meetings
      description: Retrieve a list of meetings for the authenticated user
      tags:
        - Meetings
      parameters:
        - name: status
          in: query
          description: Filter by meeting status
          schema:
            type: string
            enum: [scheduled, completed, failed]
        - name: limit
          in: query
          description: Maximum number of meetings to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of meetings to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of meetings
          content:
            application/json:
              schema:
                type: object
                properties:
                  meetings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Meeting'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create meeting
      description: Create a new meeting record
      tags:
        - Meetings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMeetingRequest'
      responses:
        '201':
          description: Meeting created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Meeting:
      type: object
      properties:
        id:
          type: string
          description: Unique meeting identifier
        title:
          type: string
          description: Meeting title
        startTime:
          type: string
          format: date-time
          description: Meeting start time
        endTime:
          type: string
          format: date-time
          description: Meeting end time
        platform:
          type: string
          enum: [zoom, teams, google-meet]
          description: Meeting platform
        meetingUrl:
          type: string
          format: uri
          description: Meeting URL
        status:
          type: string
          enum: [scheduled, completed, failed]
          description: Meeting status
        attendees:
          type: array
          items:
            type: string
          description: List of attendee emails
        transcript:
          type: string
          description: Meeting transcript
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateMeetingRequest:
      type: object
      required:
        - title
        - startTime
        - endTime
        - platform
      properties:
        title:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        platform:
          type: string
          enum: [zoom, teams, google-meet]
        meetingUrl:
          type: string
          format: uri
        attendees:
          type: array
          items:
            type: string
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        hasMore:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...